+++ lib/model/shopping_cart_model.dart
@@ -3,8 +3,8 @@ import 'package:chifa_el_meson/model/cart_item_model.dart';
 class ShoppingCart {
   List<CartItem> cartItems = [];
   ShoppingCart();
-  double get shoppingCartPrice =>
+  int get shoppingCartPrice =>
       cartItems.fold(0, (total, cartItem) => total + cartItem.cartItemPrice);
-  double get shoppingCartDiscount =>
+  int get shoppingCartDiscount =>
       cartItems.fold(0, (total, cartItem) => total + cartItem.cartItemDiscount);
 }
